Eddited button layout, moved search bar, added confirmation before deleting...

See merge request I408362/sioux-project-s3!9

+++ src/Sioux/MainController.java
@@ -13,9 +13,11 @@ import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class MainController {
     private final AppointmentController appointmentController;
@@ -26,7 +28,7 @@ public class MainController {
 
     //FXML vars
     @FXML
-    private ListView<Event> lvAllAppointments;
+    private ListView<Event> lvVisitorAppointments;
     @FXML
     private TextField tfVisitorName;
     @FXML
@@ -57,6 +59,7 @@ public class MainController {
     @FXML
     Button btnAddVisitor;
 
+
     public MainController(){
         eventList = new ArrayList<>();
         appointmentController = new AppointmentController();
@@ -65,7 +68,7 @@ public class MainController {
 
     public void initialize() {
         getAllAppointments();
-        lvAllAppointments.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
+        lvVisitorAppointments.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
 
         //Visitor page
         getAllVisitors();
@@ -76,12 +79,12 @@ public class MainController {
     private void getAllAppointments(){
         eventList = appointmentController.getEvents();
         for (Event event : eventList) {
-            lvAllAppointments.getItems().add(event);
+            lvVisitorAppointments.getItems().add(event);
         }
     }
 
     public void viewSelectedAppointment(){
-            Event selectedEvent = lvAllAppointments.getSelectionModel().getSelectedItem();
+            Event selectedEvent = lvVisitorAppointments.getSelectionModel().getSelectedItem();
             appointmentController.getEventById(selectedEvent.getId());
             tfVisitorName.setText(selectedEvent.getVisitor());
             dpAppointmentDate.setValue(selectedEvent.getStart());
@@ -89,39 +92,39 @@ public class MainController {
     }
 
     public void saveAppointment(){
-        Event selectedEvent = lvAllAppointments.getSelectionModel().getSelectedItem();
+        Event selectedEvent = lvVisitorAppointments.getSelectionModel().getSelectedItem();
         selectedEvent.setVisitor(tfVisitorName.getText());
         selectedEvent.setSubject(tfNotes.getText());
         selectedEvent.setStart(dpAppointmentDate.getValue());
         appointmentController.editEvent(selectedEvent);
-        lvAllAppointments.refresh();
+        lvVisitorAppointments.refresh();
     }
 
     public void searchForAppointment(){
         List<Event> filteredList;
         if (!tfSearchAppointments.getText().isEmpty() && dpDateSearch.getValue() == null){
             filteredList = appointmentController.searchForEventString(tfSearchAppointments.getText());
-            lvAllAppointments.getItems().removeAll(lvAllAppointments.getItems());
+            lvVisitorAppointments.getItems().removeAll(lvVisitorAppointments.getItems());
             for (Event e : filteredList){
-                lvAllAppointments.getItems().add(e);
+                lvVisitorAppointments.getItems().add(e);
             }
         }
         else if (tfSearchAppointments.getText().isEmpty() && dpDateSearch.getValue() != null){
             filteredList = appointmentController.searchForEventDate(dpDateSearch.getValue());
-            lvAllAppointments.getItems().removeAll(lvAllAppointments.getItems());
+            lvVisitorAppointments.getItems().removeAll(lvVisitorAppointments.getItems());
             for (Event e: filteredList){
-                lvAllAppointments.getItems().add(e);
+                lvVisitorAppointments.getItems().add(e);
             }
         }
         else if (!tfSearchAppointments.getText().isEmpty() && dpDateSearch.getValue() != null){
             filteredList = appointmentController.searchEventStringDate(tfSearchAppointments.getText(), dpDateSearch.getValue());
-            lvAllAppointments.getItems().removeAll(lvAllAppointments.getItems());
+            lvVisitorAppointments.getItems().removeAll(lvVisitorAppointments.getItems());
             for (Event e: filteredList){
-                lvAllAppointments.getItems().add(e);
+                lvVisitorAppointments.getItems().add(e);
             }
         }
         else{
-            lvAllAppointments.getItems().removeAll(lvAllAppointments.getItems());
+            lvVisitorAppointments.getItems().removeAll(lvVisitorAppointments.getItems());
             getAllAppointments();
         }
     }
@@ -197,14 +200,23 @@ public class MainController {
         }
     }
     public void deleteVisitor(){
-        Visitor visitorToDelete = visitorController.deleteVisitor(selectedVisitor.getVisitorID());
-        if(visitorToDelete != null){
-            lvAllVisitors.getItems().remove(visitorToDelete);
-            visitorController.getVisitorList().remove(visitorToDelete);
-            clearInfo();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation Dialog");
+        alert.setHeaderText(null);
+        alert.setContentText("Are you sure you want to delete the user?");
+        Optional <ButtonType> action = alert.showAndWait();
+            if(action.get() == ButtonType.OK) {
+                Visitor visitorToDelete = visitorController.deleteVisitor(selectedVisitor.getVisitorID());
+                if (visitorToDelete != null) {
+                    lvAllVisitors.getItems().remove(visitorToDelete);
+                    visitorController.getVisitorList().remove(visitorToDelete);
+                    clearInfo();
+                }
+                lvAllVisitors.refresh();
+            }
         }
-        lvAllVisitors.refresh();
-    }
+
+
     public void searchVisitorByName(){
         lvAllVisitors.getItems().clear();
         clearInfo();
